Loading the first found plugin.
Previously, the if there were two plugins using the same name the last one
found was the one that was loaded. This behaviour has been changed such that
the first plugin found is the one that is loaded.

The Gem#find_files method searches the load path first and then gems sorted by
their names / version numbers. Load the first found plugin favors the load
path and newer gem releases.